tic and vis refactor
