+++ aeon/visualisation/results/__init__.py
@@ -1,0 +1,1 @@
+"""Plotting tools for estimator results."""
