+++ aeon/visualisation/results/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Results plotting tests."""
