+++ aeon/visualisation/series/__init__.py
@@ -1,0 +1,1 @@
+"""Plotting  tools for time series data."""
